+++ cypress/e2e/spec.cy.js
@@ -9,7 +9,7 @@ it("Interrupts py5 draws when stop button clicked", () => {
   cy.visit(baseUrl);
   cy.get("div[class=cm-content]").invoke(
     "text",
-    "import py5\ndef setup():\n\tsize(400, 400)\ndef draw():\n\tbackground(255)\nrun()",
+    "import py5\ndef setup():\n\tpy5.size(400, 400)\ndef draw():\n\tpy5.background(255)\npy5.run_sketch()",
   );
   cy.get(".btn--run").click();